Moving _get_pb_property_value from bigtable into core.
Doing this so that we can factor out `_has_field` and use it across
packages. This is because `HasField()` works differently in
`proto3` and we use this behavior from time to time.
