+++ gcloud/test__helpers.py
@@ -526,6 +526,72 @@ class Test__datetime_to_pb_timestamp(unittest2.TestCase):
         self.assertEqual(self._callFUT(dt_stamp), timestamp)
 
 
+class Test__has_field(unittest2.TestCase):
+
+    def _callFUT(self, message_pb, property_name):
+        from gcloud._helpers import _has_field
+        return _has_field(message_pb, property_name)
+
+    def test_unset_simple_field(self):
+        from gcloud.bigtable._generated import (
+            bigtable_cluster_data_pb2 as data_pb2)
+        cluster_pb = data_pb2.Cluster()
+        self.assertFalse(self._callFUT(cluster_pb, 'serve_nodes'))
+
+    def test_set_simple_field(self):
+        from gcloud.bigtable._generated import (
+            bigtable_cluster_data_pb2 as data_pb2)
+        serve_nodes = 119
+        cluster_pb = data_pb2.Cluster(serve_nodes=serve_nodes)
+        self.assertTrue(self._callFUT(cluster_pb, 'serve_nodes'))
+
+    def test_unset_message_field(self):
+        from gcloud.bigtable._generated import (
+            bigtable_cluster_data_pb2 as data_pb2)
+        from gcloud.bigtable._generated.timestamp_pb2 import Timestamp
+        cluster_pb = data_pb2.Cluster()
+        # Check that no fields are attached to the protobuf message.
+        self.assertEqual(len(cluster_pb._fields), 0)
+        self.assertEqual(cluster_pb.delete_time, Timestamp())
+        # Check that the field is now attached to the protobuf message,
+        # even though the value is unset.
+        self.assertEqual([field.name for field in cluster_pb._fields.keys()],
+                         ['delete_time'])
+        # Make sure has field is still False.
+        self.assertFalse(self._callFUT(cluster_pb, 'delete_time'))
+
+    def test_set_message_field(self):
+        from gcloud.bigtable._generated import (
+            bigtable_cluster_data_pb2 as data_pb2)
+        from gcloud.bigtable._generated.timestamp_pb2 import Timestamp
+
+        timestamp = Timestamp()
+        cluster_pb = data_pb2.Cluster(delete_time=timestamp)
+        self.assertTrue(self._callFUT(cluster_pb, 'delete_time'))
+
+
+class Test__get_pb_property_value(unittest2.TestCase):
+
+    def _callFUT(self, message_pb, property_name):
+        from gcloud._helpers import _get_pb_property_value
+        return _get_pb_property_value(message_pb, property_name)
+
+    def test_it(self):
+        from gcloud.bigtable._generated import (
+            bigtable_cluster_data_pb2 as data_pb2)
+        serve_nodes = 119
+        cluster_pb = data_pb2.Cluster(serve_nodes=serve_nodes)
+        result = self._callFUT(cluster_pb, 'serve_nodes')
+        self.assertEqual(result, serve_nodes)
+
+    def test_with_value_unset_on_pb(self):
+        from gcloud.bigtable._generated import (
+            bigtable_cluster_data_pb2 as data_pb2)
+        cluster_pb = data_pb2.Cluster()
+        with self.assertRaises(ValueError):
+            self._callFUT(cluster_pb, 'serve_nodes')
+
+
 class _AppIdentity(object):
 
     def __init__(self, app_id):
